LayoutTests/imported/w3c/web-platform-tests/dom/events/Event-constructors-expected.txt


FAIL Event constructors assert_throws: function "function () {
    new Event()
  }" did not throw
PASS Event constructors 1 
FAIL Event constructors 2 assert_equals: expected (boolean) false but got (undefined) undefined
FAIL Event constructors 3 assert_equals: expected (boolean) false but got (undefined) undefined
FAIL Event constructors 4 Event is not a function. (In 'Event("test")', 'Event' is an instance of EventConstructor)
PASS Event constructors 5 
PASS Event constructors 6 
PASS Event constructors 7 
PASS Event constructors 8 
PASS Event constructors 9 
PASS Event constructors 10 
PASS Event constructors 11 
PASS Event constructors 12 
